silence %lu vs size_t warning by casting to (unsigned long)
(cherry picked from commit 1dc8208a89d9fc4ca5e97addb2bd76c1b9c48ea1)

+++ lib/dns/tests/rdata_test.c
@@ -247,7 +247,7 @@ check_wire_ok_single(const wire_ok_t *wire_ok, dns_rdataclass_t rdclass,
 				   wire_ok->lineno,
 				   dns_test_tohex(wire_ok->data, wire_ok->len,
 						  hex, sizeof(hex)),
-				   wire_ok->len);
+				   (unsigned long)wire_ok->len);
 	} else {
 		ATF_REQUIRE_MSG(result != ISC_R_SUCCESS,
 				"line %d: %s (%lu): "
@@ -255,7 +255,7 @@ check_wire_ok_single(const wire_ok_t *wire_ok, dns_rdataclass_t rdclass,
 				wire_ok->lineno,
 				dns_test_tohex(wire_ok->data, wire_ok->len,
 					       hex, sizeof(hex)),
-				wire_ok->len);
+				(unsigned long)wire_ok->len);
 	}
 	/*
 	 * If data was parsed correctly, perform two-way conversion checks
